Tried to reduce the stack usage of the jitty rewriter somewhat.
The jitty rewriter appears to use at least 500 bytes on the stack to rewrite a single
nested function symbol. When the rewrite rules are complex this can even
be more. With this commit this usage of the stack is reduced. Only if
there are rewrite rules with many variables this may have a substantial
effect. Otherwise the effect is still minor

It is not clear why the rewriter is using so much stack space, but it
seems that as the compiler using aggressive inlining as well as tail
recursion elimimination, many variables of auxiliary functions must
be stored in the main body of recursive function calls. Still it is not
really clear whether this is the real cause. It appears to me that
the use of 100 to 200 bytes is defendable, not 500.

To understand this properly, this requires more investigation.