Added a special function to rewrite constants in the jitty rewriter.
Constants and applications were treated in one function. This meant
that there were many unnessary declarations of data structures when
rewriting a constant. As rewriting constants is a common operation,
an optimized function has now been added to rewrite constants. For
the 1394 example this gives a (disappointing) 5% speedup in rewriting.