Experiment changing rewrite rules in the lineariser.
By replacing the rules in the lineariser to test whether
a state variable is equal to a certain value, a substantial
speedup appears to be possible.

Each generated summand has generally conditions of the shape:

    s1 == 4  && s7 == 9 && ....

The rewriter spends a substantial amount of time to construct and
rewrite the 4 and the 9. Therefore, experimentally, the generation
of the conditions have been replaced by:

    is_equal4(s1) && is_equal9(s7) &&

with the appropriate rewrite rules for is_equal4 and is_equal9.

It appears that this has a noticable effect on state space generation,
where I observed experimentally reductions of over 50% of the running
time.