More sensible comments
A few old comments that still referred to the version of
Groote/Jansen/Keiren/Wijs 2017 have been replaced by comments that belong
to the current version.

Also, when transitions are duplicated often enough to endanger the time
complexity, a warning message is printed (and the time budget is increased
so the algorithm can run, although with less strict timing checks).