Replaced closed data expressions in an lps by variables to avoid that they are rewritten very often.
Some experiments last week indicated that a substantial amount of time
(between 60% and 20%, measures vary and are not pure) of the total time
to generate a state space was spent in rewriting natural number in
expressions s == n where n is a number. These expressions occur very
often in conditions. There are several ways to circumvent this.

This commit contains a local solution, which will not change a linear
process. The (non activated) changes in the lineariser have been
reverted in this commit.

In this commit before starting to generate a state space, all constant
expressions t in the lps are replaced by variable x. Subsequently,
the assignment x:=normalform(t) is added to the substitution that is
used to generate the state space. This has as automatic effect that
the locally introduced variables are replaced by the normal form
when needed, avoiding rewriting the closed expression.

In measurements with 1394-fin this brought a 20% to 30% speedup of
a labelled transition system. A further speed improvement is possible
by not using equality to test a state (e.g. s == 5), but to use
dedicated functions such as equal_to5(s). These functions are simpler
and used a lot.

There is also a small change in the rewrite rules for pos, to reduce
the number of terms that are constructed. I.e.

  @cDub(b,p) == @cDub(c,q)   =   b==c && p==q

is replaced by

  @cDub(b,p) == @cDub(b,q)   =   p==q
  @cDub(true,p) == @cDub(false,q)   =   false
  @cDub(false,p) == @cDub(true,q)   =   false

This avoids the construction of b==c and the conjunction. This has
little effect on the jittyc rewriter, which avoids such constructions,
but the jitty rewriter becomes noticable faster by this change also.