Added machine_numbers as an instance of a data_expression.
A machine number is an aterm_int. A data expression can now also be an
aterm_int. This means that a data_expression is not a superclass anymore
of an aterm_appl, but of an aterm.
It turns out that at quite a number of places the toolset uses that
data_expressions are an aterm_appl. This is generally incorrect, and
has been repaired at quite a number of places, requiring a down_cast
to an application in general.

Most tests succeed. Some tests related to jittyc, appear to have
some residual problems.