Reading labels in .aut files erroneously remove spaces.
In bugreport #1492 it is indicated that labels in .aut files
loose their spaces when converted:

With input
des (0,1,2)
(0,"x y",1)

the result (2.aut) is

des (0,1,2)
(0,"xy",1)

The reason for this is the string streaming interface which uses

is >> skipws >> ch;

to indicate that a flag must be set to skip spaces. This flag is
permanent, in the sense that it also cause spaces to be skipped in
a subsequent

is >> ch;

In particular, although the notation suggests it, skipws is NOT a
filter. By explicitly switching space skipping off, using

is >> noskipws >> ch;

the problem is resolved.

The notation of skipws and noskipws is utterly confusing, and in my
opinion a bad language design inviting errors in programs.