Error message in state space generation erroneously showed internal variables.
When state space generation would not work, the error
does not rewrite to true or false in the condition getstate(s1_TrafficLight) == @rewr_var

could show up. The variable @rewr_var is internal and should not be
shown. It represents a rewritten constant. Before printing the message
the constant is substituted for the variable to repair this problem.