The lineariser did choose type names that overlapped with sort aliases.
When defining

sort Stack = List(Nat);

and linearising with the -lstack option the lineariser would create
a type Stack, not checking whether this name was already used. The
reason for this was that it was forgotten to add the names for sort
aliases to the protected names in the unique name generator.