[mcrl2ide] Changed the format and influence of the project file
The new format is an XML format for better usability and scalability. If a project is opened with the old project file format, it will be converted to the new format and the user will be notified of this. This conversion uses about 60 lines of code, which should be removed in the future (about 1 year).

The project file now also contains the names of the properties of the project. This enforces an order on the list of properties.