Set the size of a stack explicitly at compile time for APPLE.
Using the linker option -Wl,-stack_size,${MCRL2_OSX_STACK_SIZE}
the size of the stack on a mac is set to 4Gig. There is a variable
that shows up in ccmake namely MCRL2_OSX_STACK_SIZE
that can be set manually to set the size of the stack.

The change has no influence on Linux or Windows.