Adapted the datatype declarations.
Adapted the data type declarations to include explicit designators
internal and external to indicate whether the designators are used
in mCRL2 specifications or just internal. Also added the keywords
defined_by_code and defined_by_rewrite_rules to make a clear distinction
between data types for which explicit c++ implementations will be
provided and implementations that are based on rewrite rules.

As it stands there are problems generating c++ code from the scripts.
So, do not expect this to work.

Furthermore, as there are no implementations for the machine number
functions, most toolset operations do not work.