Added the code that takes care that a data specification knows about predefined c++ implementations of some data types.
This commit does compile, but it is not expected to work correctly
because the c++ implemented functions are not yet linked to the
functions added to the data type.
Furthermore, the rewriters have not been changed at all, and do not know
about the existence of the c++ implemented functions yet.