The first state space has been generated using "machine numbers".
The statespace of abp.aut is generated using machine numbers.

There are at least two problems at the moment.
The first one is that the pretty printer has to be adapted to machine
numbers.
The second one is that for some reason the automatically generated
well typedness checks do not include machine numbers for whatever
reason.