[mcrl2ide] changed how up to date output file checking is done
Now the paths of the input file(s) and output file are added as properties to the subprocesses to be used for up to date checking, meaning that now all filepaths used in a process (input and output for tools and up to date checking) are at the creation of a process.
This makes the changes of the previous commit unnecessary, so those changes haven (mostly) been reverted.