Substantial progress with the 64 bit data implementation
Made both compilers work with the new 64 bit numbers.
Removed a large number of mistakes, helped by the testframework.

There are still a number of failing tests, often due to the fact
that these rely on the use of the cDub constructor.

Furthermore, some tests cannot handle the new constructors
succ and zero/one, and appear to rely on te old constructors.

Initial tests on performance are not very conclusive. It seems that
minor details have large effects on the performance.

The largest activity that still needs to be undertaken is
to implement mod/div and sqrt functions on multiple machine
digits simultaneously.