Adapted the typechecker and jittyc to machine numbers.
The match trees have been adapted to also include machine numbers.

Some of the standard functions are now imported automatically into
the typechecker, instead of being declared explicitly. This
substantially reduces the length of the code. This resolved
a problem with non declared functions.

As it stands bags, fbags, sets and fsets have some problems
when including their standard functions automatically. This needs
to be investigated further.