Investigated the performance of the new 64-bit rewriter.
The new rewriter can be up to 10 times faster for very calculation
intensive tasks than the old rewriter. For ordinary rewriting the
performance can drop with 25% most likely due to the fact that we have
to deal with one extra term type, namely a machine number. It does not
appear to be easy to overcome this, as at crucial points one extra case
has to be dealt with to handle machine numbers.