Implemented operations on multiple words via the boost multiprecision library
The boost multiprecision library exactly provided algorithms 128 and 256
bit operations for mod/div and square root. In order to avoid that the
boost include files must be available for the compiling rewriter its use
has been hidden in a .cpp file.

Added tests with calculations of large numbers with up to 100 decimal
digits. They all appeared to work smoothlessly, which I believe is a
tribute to the deep analysis of the rewrite rules before putting them in
the toolset.