Increased the aterm_binary_io format number.
The new data types are incompatible with the old data types.
Therefore, I increased the number of the aterm_binary_io format
to reflect this change.

Furthermore, extensively tested the new numbers.
The conclusion is that by and large the new format is faster, sometimes
up to 10 times when heavy calculations are involved. However,
with the jitty rewriter, there are cases where new rewriter is 20%
slower than before. This is most likely due to the fact that
at many places it must be checked whether a data_expression is not an
aterm_int.