Slightly adapted the definitions for multiplication of positive and natural numbers.
The definitions were brought in line with the description in the paper.
This in particular guaranteed that the complexity of two numbers with
n and m digits have complexity Theta(n*m) and not more. In particular
not Theta(n*m+n*n) in case n is the longest and m the shortest
multiplicand, which can happen when multiplyingin naively.