Optimised polymorphic function constructor for non templated types
Calling functions such as sort_real::plus used to be very expensive,
because a new function symbol was constructed every time. However, the
number of possible results is limited, so they might as well be stored
in static variables. This also improves the performance of related
recogniser functions such as sort_real::is_plus_application.

Constructor functions for templated types have not been changed, since
the result will always depend on the template type, which is not known a
priori.