Made a new implementation of printing operators.
The old implementation was trying to minimize the amount of parentheses, but did not have enough information for that. Now printing is based on the three functions precedence, is_left_associative and is_right_associative. For the last two functions it may need to be checked if all operators are registered correctly.

Fixes #1483