Only one version of coroutines + timings in ltsconvert
The GJKW and the DNJ bisimulation minimisation algorithm used two different
versions of coroutines.  This commit makes the two algorithms use the same
version, namely the one that does not require coroutines to be methods of
a class.  (In tests, I couldn't find significant differences in timing or
memory use.)  Using only one version of coroutines should make the code
easier to maintain.

Additionally, I added more detailed timing measurements to ltsconvert, which
can be used to measure the runtime of bisimulation minimisation algorithms
without the time for file input/output, which is (hopefully) more exact.