Added a variant of enumeration that uses a callback function.
This should fix issue #1511, in the sense that the user can interrupt enumeration whenever a solution is found. The callback mechanism doesn't work well in combination with iterators, so the iterator interface is left unchanged. IMO the callback interface is preferable over the iterator interface.

Note that it is possible to make the iterator interface behave similarly. But this requires storing very detailed state information of enumeration, such that it can be stopped and resumed. This uglifies the code, and it comes with a performance penalty.

Added a small optimization to enumeration that improves the runtime of 1394 with a couple of percent.

Updated the documentation.

Fixes #1511