Simplified strong bisimilarity
When one calculates strong bisimilarity, there is no need to look for
inert transitions.  So I added a simple test that skips looking for inert
transitions if a block does not have non-bottom states.  This may also
accelerate branching bisimilarity in some cases.

Also found a way to simplify the time complexity check for finding blue
bottom states in refine().