Applied a fix for alphabet reduction of pCRL expressions.
A missing call to apply_allow has been added for pCRL expressions. This call takes care of wrapping the result of push_allow inside an allow if needed. For pCRL expressions this didn't happen. This has probably something to do with the special handling of pCRL expressions that has been added later.

Fixes #1519