Made esthetical/performance changes to aterm.h and aterm_int.h and added doxygen comments that were forgotten.
Some changes have been made to reduce the number of increments and
decrements of reference counts, and to prevent explicit initialisation
of terms that are later overwritten. These are low level operations
and it is possible that the compiler manages to optimise these away,
but if not a definitive performance improvement can be expected.