Removed the aterm_appl(const aterm&) constructor
Removed the aterm_appl(const aterm&) constructor and replaced
it with down_cast<aterm_appl> whereever applicable. This led to
sometimes uglier code, but the down_cast construction is more
efficient. At some places introduced the use of functions that
yield aterm_appl's instead of aterm's that then have to be recast.