Disabled a delete on the mac as Qt appears to be faulty.
Disabled delete m_application on a mac as it leads to a
crash in what appears to be the internal data structures within
ltsgraph. The error does not occur in diagraphica or ltsview.

Disabling this delete is awkward but the only solution with which
I can come up for now.

It might be that the error is caused by typical handling of
threads in ltsgraph, but I cannot find any evidence for this.

This resolves bug report #1517.