Resolved bug #1497 where time tags were incorrect in an lts.
Resolved #1497 which reports that a timed transition a@3
shows itself as a@2. The reason for this was that each multi-action
got a sequence number based on the actions and not on the time tag.
An dedicated indexed_set construction has been made such that
actions and time tags are now the base for sequence numbers.

This bug was the result of an incomplete implementation of time
in transition systems.