Cleaned up the code for enumeration caching.
Also did an experiment to cache action and next state computations (called "transition argument caching" in lps2lts), using two additional maps. This turns out to be too costly. So a more clever solution is needed that uses only one map. Unfortunately the pseudo code for "transition argument caching" is incomplete and unreadable, so this needs to be redone from scratch.