Made a new implementation for the function find_representative.
The function find_representative should return a unique representative of the reachable states from a vertex. The existing implementation (that was borrowed from lps2lts) didn't do this. The new version is a recursive implementation of Tarjan. It should be replaced by an iterative implementation, so that the algorithm can be stopped as soon as an SCC was found.