Mutable indexed substitutions maintain rhs vars carefully.
Up till now a mutable indexed substitution would approximate
the variables occurring in the rhs. This would lead to too many
renamings of bound variables in lambda expressions.
The indexed substitutions have now been adapted to carefully maintain
the variables in the rhs. This turns out to be used only in lambda
expressions, but it should have a noticable reduction of renamings.