Added an initial version of divergence detection.
Divergence detection is implemented as a recursive application of state space generation. It turned out that there were two technical problems with the recursion: the substitution needs to be reset more often, and the identifier generator needs to be cleared less often. This has been solved by introducing a boolean attribute recursive to the explorer class. A less intrusive solution would be preferable here.

Currently the results of divergence detection are not being cached. Also the loops are not stored. Both of these look like straightforward additions.