Revert a recent change in highway search.
After some more thought it turns out that within the current design it is not straightforward to follow the original specification of highway search. All outgoing transitions of the states in the todo set need to be examined, in order to check properties like deadlock and nondeterminism. But not all of these transitions and corresponding target states should end up in the resulting state space. To fix this afterwards, states need to be renumbered, and transitions need to be filtered. This operation is too cumbersome, so an alternative design is needed to exactly follow the original specification of highway search.