Applied a fix to the partial solve optimization of pbessolve.
The problem was that the todo set is used to represent the unexplored nodes. But there is interference with the reset function of the todo set that may remove some of those nodes. The problem was detected using random testing.

Also added an is_defined method to structure_graph. It should return true as a precondition for solving the structure graph.