Disable pruning of the todo set for strategies less than 3.
For strategies less than 3 pruning of the todo set may result in undefined nodes in the resulting graph. Recently an assertion for this has been added, which was triggered in the random tests. Pruning of the todo set has the side effect that certain nodes may already be solved. This is unexpected and undesirable for strategies <= 2, so pruning has been disabled for these cases.