Added stochastic state space generation.
It has been implemented using a lot of code duplication. This duplication can be removed once we move to c++ 17.

The generated state spaces are currently printed to standard out. Saving the generated state spaces in different LTS formats should be straightforward, but the LTS data types don't have the proper interfaces for that yet.