Applied a fix for pruning the todo set.
When pruning of the todo set is enabled, the insertion of new elements to the todo set needs to be slightly altered. Elements that have been removed from tod (and are already present in discovered) may need to be inserted in todo again. This has now been implemented.

Updated the documentation.