Fixed: linear inequalities are sometimes not constructed in normal form
This occurred when the factor of the first variable was a negative
number other than -1. The normal form requires that factor to be either
-1 or 1. Due to a bug in one of the constructors, the linear inequality
was divided by the negative inverse of the first factor.

This can lead to serious issues when using Fourier-Motzkin elimination:
the results were plain wrong.

This bug is now solved. Furthermore, I improved the implementation, by
using the abs function.

I added a regression test.