Apply pbesstategraph as post processing in random SRF tests.
Rewriting a PBES into SRF format has the effect that about 50% of the randomly generated PBESs can no longer be solved by pbessolve. Applying pbesstategraph as a post processing step solves this problem.

A check for well typedness of the generated PBESs in SRF format has been added to the random tests.