Made a new implementation of divergence detection.
The existing implementation was wrong. It turns out that it is not possible to express divergence in terms of the event points discover_state, examine_transition, start_state and finish_state.

Two new algorithms have been implemented, generate_state_space_dfs_recursive and generate_state_space_dfs_iterative, that can be used for cycle detection (and hence for divergence detection). They are quite close to how these algorithms are implemented in boost. Currently these algorithms are only available in the C++-17 version of the code. The iterative version is untested. Stochastic specifications are not yet supported.

Updated the documentation.