Action labels need not have transitions
Originally, the code for branching bisimulation minimisation contained an assertion
to ensure that every action label has at least one transition.  If there are many
action labels without transition at all, then the O(m log n) time bound cannot be
guaranteed.  The assertion failed in one example... and I replaced it by a warning
message -- it is only the timing, not the correctness of the resulting minimized
LTS, that is affected.