Repaired bug #1536.
This specification requires that the sum operator is
distributed over a distribution, i.e. in a specification such as

  sum b:Bool. dist b1:Bool[if(b,1/2,if(b1,1/3,2/3))].a(b,b2).delta

The linear process requires that the dist operator precedes the sum
operator. In general it is not easily possible to reverse this order
(I believe that the semantics is not even properly understood if
the involved types are for instance reals instead of booleans), but
in the case of finite datatypes a transformation is possible.

This has been implemented.

There are a few testcases that need to be added which will happen in the
next commit.