Fixed #1539. Stochastic linearisation.
When adding a process equation the parameters are minimized to
those occurring in the body of a process. When constructing the
initial distribution, this was not taken into account, leading
to a mismatch between the arguments of a process and the number
of its parameters. This problem was found using a random test
of generating a state space with probabilities.

The problem is resolved and a test case has been added.