Simplified version of branching bisimulation algorithm
This is the radically simplified version of the branching bisimulation
algorithm: basically we could make do without postprocessing of new bottom
states altogether.

Comments have been updated to refer to the current version of Algorithm 1
in the paper being written (but comments referring to Algorithm 2 have not
yet been updated to the correct line numbers).