Improved speed of lts algorithms
Replace a map by an unordered_map in the tau-loop removal
and unreachability algorithms for sccs. On a randomly generated
transition system 10M states and 16M transitions, of which
1.5M are inert transitions, this has a very substantial effect on
performance, namely 40% on reachability and 50% on tau-loop removal.
It is noteworthy that these times dominate the reduction of
branching-bisimulation reductions completely. Applying
branching bisimulation is neglectable, which suggests that
the SCC and reachability algorithms dan be improved much further.