Bug fix: spurious error message about empty vectors
If a LTS has no transitions, the action_block array is empty.  Then front() and
back() are undefined, and begin() and end() cannot be dereferenced.  This
sometimes led to spurious errors in MSVC, which checks for such conditions in
its internal debugging mode.

This patch fixes the bug by making the action_block array a dynamically-
allocated plain array (and some checks) insted of a std::vector.