[mcrl2ide] Implemented the ability to do equivalence checking
A new type of property ccan be created under "add property", namely an equivalence property. Here one chooses an equivalence and enters an initial process. When this property is run, the current specification is compared with the current specification in which the initial process is replaced by the initial process defined in the property.