Optimised handling transitions further
In the modified files handling transitions is now done via
indexed ordered vectors, instead of using vectors of vectors.
This leads to a speed up of at least a factor of two, and
the use of vectors of vectors is already much faster than
multimaps.

More improvements are possible, but I believe that the
major steps have now been made.