Added a case for process instantiations in the push operator of alphabet reduction.
Alphabet reduction starts by applying the push operator to the initial state. This operator was skipping process instantiations that do not appear in the context of a block or an allow operator. Now such process instantiations are wrapped in an empty block operator. This solves the use case supplied by Verum.

Adresses #1463