[mcrl2ide] Changed file usage for parsing properties
When parsing a property via the parse button in the addeditpropertydialog, the propertyfile is now saved in the temporary folder instead of the properties folder. A property is only saved in the properties folder via the save button in the same dialog. This removes the need of the method deleteUnlistedPropertyFiles, which had some unwanted behaviour (deleting property files that could not be read).