[mcrl2ide] Fixed and simplified the code for deleting properties
It no longer refuses to remove the property from the project if it could not remove the property file. (Actually, it did remove it from the project file but not from the propertiesdock due to a recent commit.)
Also removed the error information box for when removing a property file failed, as I don't really see its added value anymore.