Added code to replace indexed sets.
Added code to replace indexed sets by
unordered sets. As it stands indexed sets
are more efficient both in memory and in speed,
but this may be different when replacing
unordered sets by utilities/unordered_sets
which are slightly more efficient.

The changes are put under a define and as
such are not influencing the code.