Adapted the implementation of utilities::indexed_set
Adapted the implementation of utilities::indexed_set to
reduce its memory footprint and bring it in line with that
of atermpp::indexed_set. Adapted the interface such that it
now matches that of a vector quite closely. E.g. the
function "at" and the operator [] can now be used to find
an element given its index. A separate function "index" gives
the index for a key.

There are a few residual problems with template parameters,
in particular "Equals" and the allocator. This does not influence
the current operation tough.

All tests have been run and succeed.