Made the DNJ bisimulation reduction tool the default.
The DNJ bisimulation reduction algorithm is much more memory efficient
than the GJKW bisimulation reduction algorithm which is the standard
in the toolset. The memory footprint of DNJ is comparable to the GV
algorithm.

Therefore the DNJ bisimulation algorithm is now the default algorithm,
and the GJKW as well as the GV can be invoked by using explict flags.

The only place where the DNJ algorithm is not applied is the
failure_reduction algorithm. Here a "TODO" has been added. The reason
is that DNJ did not make its partitioner public. It has been requested
to do so.

All tests succeed in debug mode.

This resolves #1462.

(cherry picked from commit 91ffd8f37fb5b763ea70ea86e909a8cb8c359d69)