Added < on reals to deal with probabilistic processes.
Added < on reals to the rewriter when generating a state space to deal with
comparisons on probabilities. Added two testcases of which one is
outcommented as it fails as depth first exploration on probabilistic statespaces
does not work properly.

This is an essential bugfix that must be ported to the release.

(cherry picked from commit 0dde6fa6e06ed04c8c069bb82a375215af51acd8)