Resolved but #1553.
The lineariser could not linearise a specification of which
the data types were not made concrete, ie. as in:

sort D;

The reason for this is that sometimes the lineariser joins summands
for instance of the shape

sum d:D. a(d).X +
         a(1).X

Into something like

sum b:Bool.if(b,true,eq(d,d0)) -> a(if(b,d,1)).X;

The condition eq(d,d0) is there to prevent enumeration of
the second summand for all valued of d, which is not necessary or
even desired.

But in some cases a default value d0 cannot be found.
In that case the tool gives an error message, which is not really
comprehensible.

The behaviour has been changed in the sense that no condition will be
added if a unique representant d0 for a domain D cannot be found.

Remarkably, this is a problem lingering in the linearisator since
the beginning and apparantly nobody really noticed the problem up till
now.

This resolves #1553.