Removed atermpp::index_set in favour of utilities::indexed_set.
Als adapted exploration.cpp such that it does not contain a variant
with unordered_maps anymore.

Moved the atermpp::index_set_test to utilities and adapted it.

Adapted ltsmin.h and besconvert.cpp which were still using
atermpp::indexed_sets.

This commit is NOT for the release.

All tests succeed, except those that already failed for other reasons.