Bug about action labels with 0 actions corrected
In particular in dpweak-bisim, it can happen that there are action
labels without any action. This led to an assertion failure during
initialisation of the branching bisimulation minimisation but should
be accepted. The current version accepts action labels without action.

Additionally, it prints a warning if the number of states is much
larger than the number of transitions, because then it is not
guaranteed that the algorithm runs in time O(m log n).