Corrected depth first search for probabilistic processes
It turned out that depth first search was never appropriately
implemented in lps2lts for probabilistic processes. This
has now been mended.

I am in favour of adding this to the release, as non functioning
depth first search for probabilistic processes is a real ommission.
However, it is a substantial commit, and it needs to be
looked into carefully. In particular it is interesting whether
there will be a noticable effect on the performance of generating
normal transition systems.

(cherry picked from commit 2618af7bcc7c1514ddcbbed5f500872b02439867)