[mcrl2ide] Added a safety measure that makes sure no duplicate property names are added to the project file
Previously it was possible to add duplicate property names to the project file by creating an equivalence property, manually editing the property file so that it becomes unreadable for mcrl2ide, reopening the project and then creating a new property with the same name. When the project is then reopened the property would be shown twice on the propertiesdock, both referring to the same file.
Also renamed a variable to be more in line with what it is.

(cherry picked from commit fb7d2fde2430d3819502971ed90d48500327ec78)