Make enumerations of stochastic variables unbounded in the state space generation.
In a state space generation the enumeration of a stochastic variable x in
dist x:Type..... was limited to
data::detail::get_enumerator_iteration_limit(). This meant that
sometimes a state space generation would terminate prematurely given an
obscure error message that an expression could not be enumerated. As in
the enumeration of variables in sums the limit for the enumeration of
such stochastic variables is now set to unlimited. This problem is
related to #1542 which was already closed, but which had not resolved
the original underlying problem.