[mcrl2ide] Postponed finishing a process early in case it was an mcrl2 parsing subprocess that failed
This is to prevent a race condition between the parsing result being stored and this result being queried after being notified that the process has finished. This can happen when there are more subprocesses after the failing parsing process.
Also renamed a variable to be more in line with what it is.