Added a small note explaining the data_equation_set
All major data types in data_specifications are vectors,
e.g. mapping_vector. The data_equations are stored in a set.
This is done to facilitate removal of double equations.
As it can be that there are many equations the use of
a vector can cause performance problems.

A small note has been made in the code to explain this.