Fix implementation of condition L
The explicit DFS stack was not managed in the right way. Therefore, the
implementation was most likely not sound. At the same time, there was a
bug in the implementation that caused many states to be fully expanded.
In a lot of cases, no reduction could be achieved at all.