Jitty and jittyc respond very differently when enumerating.
In #1557 a situation is described where jitty and jittyc
behave very diffently. The reason for this lies in the fact
that for a structured sort A = struct ....
the expression x==x with x:A can either rewrite to true
or to @equal_arguments(x,x). The last expression cannot be
simplified any further and needs to be resolved by enumerating.

Jitty used the first rule and jittyc the second.

This has been resolved by adding an extra rewrite rule
@equal_arguments(x,x)=true.

This resolves #1557.