Improve datatypes so less stack searching is necessary
Just like in Tarjan's SCC algorithm, we now store an index and a boolean
per state in a mapping. The boolean indicates whether the node is
currently in the stack. The indices corresponding to the nodes in the
stack are ordered from low to high. This helps to decide the cycle
condition much faster.