Port Sphinx generated Makefile to CMake
Also introduce a new target to clean the documentation, and make the doc target
use that target as a dependency. I had to do some trickery to make everything
work without duplicating CMake code because it is not possible to specify an
order between depencies; otherwise, the doc target could have simply depended on
first cleandoc and then fastdoc.