Make identifiers qualified in library documentation to avoid Sphinx error
Sphinx would crash (assertion error) on duplicate declarations. These
declarations are not duplicates, but Sphinx cannot know because they are not
fully qualified. This commit changes the compound XSL that is used to transform
Doxygen output into reST, so that identifiers are qualified.

Sphinx no longer crashes when building the mCRL2 website now.