Added a 'quantifier-propagate' rewriter to pbestransform
Quantifier propagation pushes quantifiers that are directly in front of
a predicate variable instantiation through that PVI. This transformation
can sometimes significantly reduce a PBES, especially when combined with
the quantifier-inside and quantifier-one-point rewriters. In some cases,
the state space becomes larger, especially when quantifier propagation
is applied repeatedly.
This 'rewriter' works a bit differently from other rewriters for
quantifiers, in the sense that it operates on a PBES as a whole. This is
necessary since it may add new equations.