Made the cmake flag MCRL2_SKIP_LONG_TESTS a non expert flag.
I noticed that many linearisation tests are not run (anymore?).
They are excluded by the cmake flag MCRL2_SKIP_LONG_TESTS. As it
was tricky to find this flag, I made it a non-expert flag, such
that when testing the toolset it is more obvious that it can
or cannot be set.

Btw. all tests succeed with this flag set.