Resolved bug #1576 which was due to a type check error.
The type checker indicated that a rewrite rule of the
shape

   f  = g

where f could have multiple types would yield a left hand side
of the shape UntypedIdentifier(f), and then it would flag this
rule as correctly typed.

This would lead to all kinds of subsequent errors in the toolset.

By strengthening typing a little bit, this left hand side is now
flagged as ambiguously typed.

I consider this as the solution for this problem.
This closes #1576.

Added a testcase to cover this.