New equivalence for ltscompare: 'coupled-sim'
This new equivalence checks the two root processes
for coupled simulation equivalence, introduced
by Parrow and Sjödin in 1992 [PS92].

The algorithm uses the game theoretic approach of
Bisping and Nestmann from 2019 [BN19].

The algorithm will return equality,
if the root processes are coupled similar.
Additionally it calculates the set R,
which can be shown with the `ltscompare` flag `--debug`.