Added an exception to warn against the problem mentioned in #1577.
In lps2ltsdeprecated the trace to an action is incorrectly printed,
as reported in #1577. This is due to the fact that when reconstructing
a trace the state vectors are regenerated. Due to the use of binders
the variables in the states are not always exactly the same. This
causes reconstruction to fail.

As this is deprecated code, I resolved the problem by throwing an
exception if it occurs.

However, state space explosion can occur when renaming variables
in binders in state vectors, and this should be avoided as much as
possible.

This resolves bug #1577.