Resolved the problem that on a mac the menubar is not selectable.
Using a somewhat nasty trick, which appears to work, the
menubar is activated on a mac directly when an application
is started from the menubar. This solution should be removed
when Qt provides a more fundamental solution.
It also requires events to be sent to applications for which
explicit permission must be given, at least in macOS Catalina.